This works with any QR reader that redirects to a website
Timelogger creates a structure of menu items that can be either selected
from a web browser or from a QR code.  Buttons and QR grids are
generated on the fly for each action with a start and stop attachment.
I will eventually be able to display streams in php GD graphs.  A stream
is a type of action that is mutually exclusive.  For example, you could
be at work, at home, or in transit, and working on a specific customer
project, but can only be in one place at a time or working on one
project at a time.  Therefore a location stream exists and a project
stream exists.